requirements.txt
scipy==1.10.1
numpy==1.23.0
matplotlib==3.7.2
opencv-python==4.8.1.78
torch==1.10.1
torchvision==0.11.2
tqdm==4.65.0
Pillow==7.2.0
h5py==2.10.0
pandas==2.0.3
pyyaml==6.0.1
requests==2.31.0
seaborn==0.12.2
ftfy
regex
clip-by-openai
einops